Refactor CMake configuration and update file paths for core components
