+++ CMakeLists.txt
@@ -1,6 +1,10 @@
 cmake_minimum_required(VERSION 3.28.3)
 project(RSA-Encryptor)
 
+# Enforce C++20 and disable compiler-specific extensions
 set(CMAKE_CXX_STANDARD 20)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+set(CMAKE_CXX_EXTENSIONS OFF)
 
 add_subdirectory(src)
+add_subdirectory(tests)
