+++ src/core/encryption.cpp
