+++ src/core/encryption.h
