+++ src/core/utility.cpp
