+++ src/core/utility.h
